Rounded box changes
- Return correct type for radius
- Move enum to header file to make it available in other translation
  units

+++ chapter_14/exercise/share/geo/box.h
@@ -49,7 +49,7 @@ public:
 	enum Drawing_mode { DRAW_STROKE, DRAW_FILL };
 	int width() const { return w; };
 	int height() const { return h; };
-	int corner_radius() const { return crv_radius; };
+	double corner_radius() const { return crv_radius; };
 	Graph_lib::Point bbox_origin() const;
 	void updateSize(int hh, int ww);
 	void updateSize(const Graph_lib::Point& o, const Graph_lib::Point& e);
